refactor: move the component to new folder
